S-01122 - “Remove support for MusicXML v1. TK-01191 - “Remove dependency on MXML v1 from IMUSANT_processing” Added IMUSANT_conversions to the main project from the music MusicXML group. Added fstream include into IMUSANT_pocessing Removed v1 files from libIMUSANT "Compile Sources" build step: - IMUSANT_XMLFile.cpp - IMUSANT_XMLReader.cpp - IMUSANT_XMLVisitor.cpp - TXML2IMUSANTVisitor.cpp
There is an issue if I remove the #include "IMUSANT_XMLReader.h" statement from IMUSANT_processing.cpp. I get a compilation problem with IMUSANT_chord.h that looks like a circular dependency, but I can't find it. XCode is trying to resolve SMARTP and failing. I think this isn a general problem, so I'm going to move on and come back to it later.